requirements.txt
phidata>=1.8.5
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
pgvector>=0.2.0
psycopg2-binary>=2.9.9
pyyaml>=6.0.0
aiohttp>=3.8.5
asyncio>=3.4.3
typing-extensions>=4.7.1 